Move type tests inside `queryLifecycle.test.tsx` into `queryLifecycle.test-d.tsx`
